Added stock controller
