+++ gamestonk_terminal/stocks/dark_pool_shorts/dps_controller.py
@@ -7,7 +7,7 @@ from datetime import datetime, timedelta
 import pandas as pd
 from prompt_toolkit.completion import NestedCompleter
 from gamestonk_terminal.rich_config import console
-from gamestonk_terminal.parent_classes import BaseController
+from gamestonk_terminal.parent_classes import StockController
 from gamestonk_terminal import feature_flags as gtff
 from gamestonk_terminal.menu import session
 from gamestonk_terminal.helper_funcs import (
@@ -18,7 +18,6 @@ from gamestonk_terminal.helper_funcs import (
     EXPORT_BOTH_RAW_DATA_AND_FIGURES,
     EXPORT_ONLY_RAW_DATA_ALLOWED,
 )
-from gamestonk_terminal.stocks import stocks_helper
 from gamestonk_terminal.stocks.dark_pool_shorts import (
     yahoofinance_view,
     stockgrid_view,
@@ -30,7 +29,7 @@ from gamestonk_terminal.stocks.dark_pool_shorts import (
 )
 
 
-class DarkPoolShortsController(BaseController):
+class DarkPoolShortsController(StockController):
     """Dark Pool Shorts Controller class"""
 
     CHOICES_COMMANDS = [
@@ -99,49 +98,6 @@ class DarkPoolShortsController(BaseController):
     volexch        short volume for ARCA,Amex,Chicago,NYSE and national exchanges[/cmds]
 {has_ticker_end}"""
         console.print(text=help_text, menu="Stocks - Dark Pool and Short data")
-
-    def call_load(self, other_args: List[str]):
-        """Process load command"""
-        parser = argparse.ArgumentParser(
-            add_help=False,
-            formatter_class=argparse.ArgumentDefaultsHelpFormatter,
-            prog="load",
-            description="Load stock ticker to perform analysis on. When the data source is 'yf', an Indian ticker can be"
-            " loaded by using '.NS' at the end, e.g. 'SBIN.NS'. See available market in"
-            " https://help.yahoo.com/kb/exchanges-data-providers-yahoo-finance-sln2310.html.",
-        )
-        parser.add_argument(
-            "-t",
-            "--ticker",
-            action="store",
-            dest="ticker",
-            required="-h" not in other_args,
-            help="Stock ticker",
-        )
-        parser.add_argument(
-            "-s",
-            "--start",
-            type=valid_date,
-            default=(datetime.now() - timedelta(days=366)).strftime("%Y-%m-%d"),
-            dest="start",
-            help="The starting date (format YYYY-MM-DD) of the stock",
-        )
-        if other_args and "-" not in other_args[0][0]:
-            other_args.insert(0, "-t")
-        ns_parser = parse_known_args_and_warn(parser, other_args)
-        if ns_parser:
-            df_stock_candidate = stocks_helper.load(
-                ns_parser.ticker,
-                ns_parser.start,
-            )
-
-            if not df_stock_candidate.empty:
-                self.stock = df_stock_candidate
-                self.start = ns_parser.start
-                if "." in ns_parser.ticker:
-                    self.ticker = ns_parser.ticker.upper().split(".")[0]
-                else:
-                    self.ticker = ns_parser.ticker.upper()
 
     def call_shorted(self, other_args: List[str]):
         """Process shorted command"""
