+++ gamestonk_terminal/stocks/fundamental_analysis/fa_controller.py
@@ -2,12 +2,11 @@
 __docformat__ = "numpy"
 
 import argparse
-from datetime import datetime, timedelta
 from typing import List
 from prompt_toolkit.completion import NestedCompleter
 from gamestonk_terminal.rich_config import console
 
-from gamestonk_terminal.parent_classes import BaseController
+from gamestonk_terminal.parent_classes import StockController
 from gamestonk_terminal.stocks.fundamental_analysis.financial_modeling_prep import (
     fmp_controller,
     fmp_view,
@@ -26,7 +25,6 @@ from gamestonk_terminal.helper_funcs import (
     EXPORT_ONLY_RAW_DATA_ALLOWED,
     parse_known_args_and_warn,
     check_positive,
-    valid_date,
 )
 from gamestonk_terminal.stocks import stocks_helper
 from gamestonk_terminal.menu import session
@@ -34,7 +32,7 @@ from gamestonk_terminal.menu import session
 # pylint: disable=inconsistent-return-statements
 
 
-class FundamentalAnalysisController(BaseController):
+class FundamentalAnalysisController(StockController):
     """Fundamental Analysis Controller class"""
 
     CHOICES_COMMANDS = [
@@ -130,74 +128,6 @@ Ticker: [/param] {self.ticker} [cmds]
         if self.ticker:
             return ["stocks", f"load {self.ticker}", "fa"]
         return []
-
-    def call_load(self, other_args: List[str]):
-        """Process load command."""
-        parser = argparse.ArgumentParser(
-            add_help=False,
-            formatter_class=argparse.ArgumentDefaultsHelpFormatter,
-            prog="load",
-            description="Load stock ticker to perform analysis on. When the data source "
-            + "is 'yf', an Indian ticker can be loaded by using '.NS' at the end, e.g. "
-            + "'SBIN.NS'. See available market in"
-            + " https://help.yahoo.com/kb/exchanges-data-providers-yahoo-finance-sln2310.html.",
-        )
-        parser.add_argument(
-            "-t",
-            "--ticker",
-            action="store",
-            dest="ticker",
-            required="-h" not in other_args,
-            help="Stock ticker",
-        )
-        parser.add_argument(
-            "-s",
-            "--start",
-            type=valid_date,
-            default=(datetime.now() - timedelta(days=366)).strftime("%Y-%m-%d"),
-            dest="start",
-            help="The starting date (format YYYY-MM-DD) of the stock",
-        )
-        parser.add_argument(
-            "-i",
-            "--interval",
-            action="store",
-            dest="interval",
-            type=int,
-            default=1440,
-            choices=stocks_helper.INTERVALS,
-            help="Intraday stock minutes",
-        )
-        parser.add_argument(
-            "--source",
-            action="store",
-            dest="source",
-            choices=stocks_helper.SOURCES,
-            default="yf",
-            help="Source of historical data.",
-        )
-        # For the case where a user uses: 'load BB'
-        if other_args and "-t" not in other_args and "-h" not in other_args:
-            other_args.insert(0, "-t")
-
-        ns_parser = parse_known_args_and_warn(
-            parser, other_args, EXPORT_ONLY_RAW_DATA_ALLOWED
-        )
-        if ns_parser:
-            df_stock_candidate = stocks_helper.load(
-                ticker=ns_parser.ticker,
-                start=ns_parser.start,
-                interval=ns_parser.interval,
-                source=ns_parser.source,
-            )
-
-            if not df_stock_candidate.empty:
-                self.start = ns_parser.start
-                self.interval = str(ns_parser.interval) + "min"
-                if "." in ns_parser.ticker:
-                    self.ticker = ns_parser.ticker.upper().split(".")[0]
-                else:
-                    self.ticker = ns_parser.ticker.upper()
 
     def call_analysis(self, other_args: List[str]):
         """Process analysis command."""
