+++ gamestonk_terminal/stocks/prediction_techniques/pred_controller.py
@@ -3,7 +3,7 @@ __docformat__ = "numpy"
 
 import argparse
 from typing import List
-from datetime import datetime, timedelta
+from datetime import datetime
 import pandas as pd
 import numpy as np
 from prompt_toolkit.completion import NestedCompleter
@@ -121,102 +121,6 @@ class PredictionTechniquesController(BaseController):
         if self.ticker:
             return ["stocks", f"load {self.ticker}", "pred"]
         return []
-
-    def call_load(self, other_args: List[str]):
-        """Process load command"""
-        parser = argparse.ArgumentParser(
-            add_help=False,
-            formatter_class=argparse.ArgumentDefaultsHelpFormatter,
-            prog="load",
-            description="Load stock ticker to perform analysis on. When the data source"
-            + " is 'yf', an Indian ticker can be loaded by using '.NS' at the end,"
-            + " e.g. 'SBIN.NS'. See available market in"
-            + " https://help.yahoo.com/kb/exchanges-data-providers-yahoo-finance-sln2310.html.",
-        )
-        parser.add_argument(
-            "-t",
-            "--ticker",
-            action="store",
-            dest="ticker",
-            required="-h" not in other_args,
-            help="Stock ticker",
-        )
-        parser.add_argument(
-            "-s",
-            "--start",
-            type=valid_date,
-            default=(datetime.now() - timedelta(days=366)).strftime("%Y-%m-%d"),
-            dest="start",
-            help="The starting date (format YYYY-MM-DD) of the stock",
-        )
-        parser.add_argument(
-            "-e",
-            "--end",
-            type=valid_date,
-            default=datetime.now().strftime("%Y-%m-%d"),
-            dest="end",
-            help="The ending date (format YYYY-MM-DD) of the stock",
-        )
-        parser.add_argument(
-            "-i",
-            "--interval",
-            action="store",
-            dest="interval",
-            type=int,
-            default=1440,
-            choices=stocks_helper.INTERVALS,
-            help="Intraday stock minutes",
-        )
-        parser.add_argument(
-            "--source",
-            action="store",
-            dest="source",
-            choices=stocks_helper.SOURCES,
-            default="yf",
-            help="Source of historical data.",
-        )
-        parser.add_argument(
-            "-p",
-            "--prepost",
-            action="store_true",
-            default=False,
-            dest="prepost",
-            help="Pre/After market hours. Only works for 'yf' source, and intraday data",
-        )
-
-        # For the case where a user uses: 'load BB'
-        if other_args and "-t" not in other_args and "-h" not in other_args:
-            other_args.insert(0, "-t")
-
-        ns_parser = parse_known_args_and_warn(
-            parser, other_args, EXPORT_ONLY_FIGURES_ALLOWED
-        )
-        if ns_parser:
-            df_stock_candidate = stocks_helper.load(
-                ns_parser.ticker,
-                ns_parser.start,
-                ns_parser.interval,
-                ns_parser.end,
-                ns_parser.prepost,
-                ns_parser.source,
-            )
-
-            if not df_stock_candidate.empty:
-                self.stock = df_stock_candidate
-                if "." in ns_parser.ticker:
-                    self.ticker = ns_parser.ticker.upper().split(".")[0]
-                else:
-                    self.ticker = ns_parser.ticker.upper()
-
-                self.start = ns_parser.start
-                self.interval = str(ns_parser.interval) + "min"
-
-                self.stock["Returns"] = self.stock["Adj Close"].pct_change()
-                self.stock["LogRet"] = np.log(self.stock["Adj Close"]) - np.log(
-                    self.stock["Adj Close"].shift(1)
-                )
-                self.stock = self.stock.rename(columns={"Adj Close": "AdjClose"})
-                self.stock = self.stock.dropna()
 
     def call_pick(self, other_args: List[str]):
         """Process pick command"""
